Replace Unittest with Pytest
This commit updates the code base to use pytest for all tests in the
project.

This commit also pushes the minimum version of Python to 3.7.
