+++ pyproject.toml
@@ -29,14 +29,10 @@ packages = [
     { include = "django_scim", from = "src" },
 ]
 
-[tool.poetry.scripts]
-test = 'scripts:run_tests'
-
 [tool.poetry.dependencies]
-python = ">=3.6"
+python = ">=3.7"
 scim2-filter-parser = ">=0.3.5"
 django = ">=3.2"
-python-dateutil = ">=2.7.3"
 
 [tool.poetry.dev-dependencies]
 mock = "^4.0.2"
@@ -44,6 +40,10 @@ tox = "^3.16.1"
 flake8 = "^3.8.3"
 toml = "^0.10.1"
 flake8-isort = "^3.0.0"
+pytest = ">=5.4.0"
+pytest-django = "4.5.2"
+coverage = "^6.4.2"
+pytest-cov = "3.0.0"
 
 [tool.black]
 line-length = 100
