+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = tests.settings
+python_files = tests/test_*.py
+django_find_project = false
+pythonpath = . src
