+++ tox.ini
@@ -4,13 +4,14 @@ envlist =
     py{37,38,39,310}-django{32}
     py{38,39,310}-django{40}
     flake8
+    coverage
 
 [gh-actions]
 python =
     3.7: py37
     3.8: py38
     3.9: py39
-    3.10: py310, flake8
+    3.10: py310
 
 [testenv]
 setenv =
@@ -28,12 +29,18 @@ deps =
 whitelist_externals = poetry
 commands =
     poetry install -v
-    poetry run test
+    poetry run pytest tests/
 
 [testenv:flake8]
 basepython =
     python3.10
-deps =
-    flake8
 commands =
-    flake8 src
+    poetry install -v
+    poetry run flake8 src
+
+[testenv:coverage]
+basepython =
+    python3.10
+commands =
+    poetry install -v
+    poetry run pytest --cov=django_scim --cov-report=xml --cov-report=term
